data

+++ app/Http/Controllers/Frontend/HomeController.php
@@ -30,9 +30,9 @@ class HomeController extends Controller
         $category = Category::paginate(20, ['*'], 'categories');
         $qualification = Qualification::paginate(20, ['*'], 'qualification');
         $jobType = JobType::paginate(20, ['*'], 'jobtypes');
-        $jobs = Add::with('getCity')->where('type' , 'jobs')->paginate(5, ['*'], 'jobs');
-        $admissions = Add::with('getCity')->where('type' , 'admissions')->paginate(5, ['*'], 'admissions');
-        $tenders = Add::with('getCity')->where('type' , 'tenders')->paginate(5, ['*'], 'tenders');
+        $jobs = Add::with('getCity')->where('type' , 'jobs')->where('status','Active')->paginate(5, ['*'], 'jobs');
+        $admissions = Add::with('getCity')->where('type' , 'admissions')->where('status','Active')->paginate(5, ['*'], 'admissions');
+        $tenders = Add::with('getCity')->where('type' , 'tenders')->where('status','Active')->paginate(5, ['*'], 'tenders');
         return view('frontend.index' , compact('newspapers' , 'cities' , 'category' , 'qualification' , 'jobType' , 'jobs' , 'tenders' , 'admissions'));
     }
 
@@ -60,7 +60,7 @@ class HomeController extends Controller
       // $date = \Carbon\Carbon::createFromFormat('d/m/Y' , $date);
       // dd($date , $request->all());
 
-        $newspaper = Add::whereHas('getNewsPaper' , function($q) use ($slug){
+        $newspaper = Add::where('status','Active')->whereHas('getNewsPaper' , function($q) use ($slug){
            $q->where('slug' , $slug); 
         })->where('type' , $type)
         ->whereDate('created_at' , $date)
@@ -71,30 +71,30 @@ class HomeController extends Controller
         return view('frontend.date-adds' , compact('paper' , 'newspaper' , 'newspapers' , 'slug' , 'date'));
     }
     public function jobType($id){
-        $newspaper = Add::where('job_type_id' , $id)->get();
+        $newspaper = Add::where('job_type_id' , $id)->where('status','Active')->get();
         $newspapers = Newspaper::all();
         return view('frontend.jobtype-adds' , compact('newspaper' , 'newspapers'));
     }
     public function cityAdds($id){
-        $newspaper = Add::where('city_id' , $id)->get();
+        $newspaper = Add::where('city_id' , $id)->where('status','Active')->get();
         $newspapers = Newspaper::all();
         return view('frontend.city-adds' , compact('newspaper' , 'newspapers'));
     }
     public function QualificationAdds($id){
-        $newspaper = Add::where('qualification_id' , $id)->get();
+        $newspaper = Add::where('qualification_id' , $id)->where('status','Active')->get();
         $newspapers = Newspaper::all();
         $qualification_id = $id;
         return view('frontend.qualification-adds' , compact('newspaper' , 'newspapers' , 'qualification_id'));
     }
     public function CategoryAdds($id){
-        $newspaper = Add::where('category_id' , $id)->get();
+        $newspaper = Add::where('category_id' , $id)->where('status','Active')->get();
         $newspapers = Newspaper::all();
         $cat_id = $id;
         return view('frontend.category-adds' , compact('newspaper' , 'newspapers' , 'cat_id'));
     }
     public function detailPage($id){
-        $add = Add::with('getCity' , 'getNewsPaper' , 'getCategory', 'getJobType' , 'getQualification')->where('id' , $id)->first();
-        $rel = Add::select('id' , 'title')->where('category_id' , $add->category_id)->take(8)->get();
+        $add = Add::with('getCity' , 'getNewsPaper' , 'getCategory', 'getJobType' , 'getQualification')->where('id' , $id)->where('status','Active')->first();
+        $rel = Add::select('id' , 'title')->where('category_id' , $add->category_id)->where('status','Active')->take(8)->get();
         return view('frontend.job_detail' , compact('add' , 'rel'));
     }
     public function uploadCv(Request $request){
@@ -165,6 +165,9 @@ class HomeController extends Controller
     }
 
     public function subscribeNews($id){
+      if(\Auth::user() == null){
+        return redirect()->guest(route('frontend.auth.register'));
+      }
       $news = subscription::where('user_id' , \Auth::user()->id)->where('newspaper_id' , $id)->first();
       if($news){
         return redirect()->back();
@@ -176,6 +179,9 @@ class HomeController extends Controller
       return redirect()->back();
     }
     public function subscribeCate($id){
+      if(\Auth::user() == null){
+        return redirect()->guest(route('frontend.auth.register'));
+      }
       $news = subscription::where('user_id' , \Auth::user()->id)->where('category_id' , $id)->first();
       if($news){
         return redirect()->back();
@@ -187,11 +193,17 @@ class HomeController extends Controller
       return redirect()->back();
     }
     public function unSubscribeCate($id){
+      if(\Auth::user() == null){
+        return redirect()->back();
+      }
       $news = subscription::where('user_id' , \Auth::user()->id)->where('category_id' , $id)->first();
       $news->delete();
       return redirect()->back();
     }
     public function subscribeQual($id){
+      if(\Auth::user() == null){
+        return redirect()->guest(route('frontend.auth.register'));
+      }
       $news = subscription::where('user_id' , \Auth::user()->id)->where('qualification_id' , $id)->first();
       if($news){
         return redirect()->back();
@@ -203,18 +215,24 @@ class HomeController extends Controller
       return redirect()->back();
     }
     public function unSubscribeQual($id){
+      if(\Auth::user() == null){
+        return redirect()->back();
+      }
        $news = subscription::where('user_id' , \Auth::user()->id)->where('qualification_id' , $id)->first();
       $news->delete();
       return redirect()->back();
     }
     public function unSubscribeNews($id){
+      if(\Auth::user() == null){
+        return redirect()->back();
+      }
        $news = subscription::where('user_id' , \Auth::user()->id)->where('newspaper_id' , $id)->first();
       $news->delete();
       return redirect()->back();
     }
 
     public function showCalenderJob(){
-      $adds = Add::where('type' , 'jobs')->groupBy(\DB::raw('Date(created_at)'))->select('created_at' , \DB::raw('count(*) as count') , 'type')->get();
+      $adds = Add::where('type' , 'jobs')->where('status','Active')->groupBy(\DB::raw('Date(created_at)'))->select('created_at' , \DB::raw('count(*) as count') , 'type')->get();
       $adds->map(function ($add) {
         $add['title'] = "Pakistan Jobs" . ": ".$add->count;
         $add['start'] = $add->created_at->toDateString();
@@ -227,7 +245,7 @@ class HomeController extends Controller
     }
 
     public function showCalenderTender(){
-      $adds = Add::where('type' , 'tenders')->groupBy(\DB::raw('Date(created_at)'))->select('created_at' , \DB::raw('count(*) as count') , 'type')->get();
+      $adds = Add::where('type' , 'tenders')->where('status','Active')->groupBy(\DB::raw('Date(created_at)'))->select('created_at' , \DB::raw('count(*) as count') , 'type')->get();
       $adds->map(function ($add) {
         $add['title'] = "Abroad Jobs" . ": ".$add->count;
         $add['start'] = $add->created_at->toDateString();
@@ -240,7 +258,7 @@ class HomeController extends Controller
     }
 
     public function showCalenderAdmission(){
-      $adds = Add::where('type' , 'admissions')->groupBy(\DB::raw('Date(created_at)'))->select('created_at' , \DB::raw('count(*) as count') , 'type')->get();
+      $adds = Add::where('type' , 'admissions')->where('status','Active')->groupBy(\DB::raw('Date(created_at)'))->select('created_at' , \DB::raw('count(*) as count') , 'type')->get();
       $adds->map(function ($add) {
         $add['title'] = "Online Jobs" . ": ".$add->count;
         $add['start'] = $add->created_at->toDateString();
@@ -262,7 +280,7 @@ class HomeController extends Controller
 
     public function showCompanyAdd($slug){
 
-      $newspaper = Add::where('company_name' , $slug)->get();
+      $newspaper = Add::where('company_name' , $slug)->where('status','Active')->get();
       $newspapers = Newspaper::all();
       return view('frontend.company-add', compact('newspaper' , 'newspapers'));
     
@@ -270,7 +288,7 @@ class HomeController extends Controller
 
     public function showDateLast($date){
 
-      $newspaper = Add::whereDate('last_date' , $date)->get();
+      $newspaper = Add::where('status','Active')->whereDate('last_date' , $date)->get();
       $newspapers = Newspaper::all();
       return view('frontend.last-date', compact('newspaper' , 'newspapers'));
     
@@ -278,7 +296,7 @@ class HomeController extends Controller
 
     public function showDateApply($date){
 
-      $newspaper = Add::whereDate('created_at' , $date)->get();
+      $newspaper = Add::where('status','Active')->whereDate('created_at' , $date)->get();
       $newspapers = Newspaper::all();
       return view('frontend.apply-date-ad', compact('newspaper' , 'newspapers'));
     
@@ -286,9 +304,9 @@ class HomeController extends Controller
 
     public function showDateType($date , $type){
 
-      $newspaper = Add::where('type' , $type)
+      $newspaper = Add::where('status','Active')->where('type' , $type)
         ->where(function ($query) use ($date){
-          $query->whereDate('created_at' , $date)->orWhereDate('last_date' , $date)->orWhereDate('apply_by' , $date);
+          $query->whereDate('created_at' , $date);
         })
         ->with('getCity')->get();
       $newspapers = Newspaper::all();
@@ -297,7 +315,7 @@ class HomeController extends Controller
     }
 
     public function NewspaperJobs($id){
-      $newspaper = Add::where('newspaper_id' , $id)->get();
+      $newspaper = Add::where('status','Active')->where('newspaper_id' , $id)->get();
       $news = Newspaper::find($id);
         $newspapers = Newspaper::all();
         return view('frontend.newspaper-adds' , compact('newspaper' , 'newspapers' , 'news'));
